moved the signup folder into pages
